updated create workspace credits table, update sqlx data.json, added customer to payment intent object

+++ api/src/migrations/from_v0/from_v0_7/from_v0_7_4.rs
@@ -489,7 +489,7 @@ pub async fn add_workspace_credit_column(
 		r#"
 		CREATE TABLE workspace_credits(
 			workspace_id UUID NOT NULL,
-			credits DECIMAL(15,6) NOT NULL DEFAULT 0,
+			credits BIGINT NOT NULL DEFAULT 0,
 			metadata JSON NOT NULL,
 			date TIMESTAMPTZ NOT NULL,
 			CONSTRAINT workspace_credits_fk_workspace_id FOREIGN KEY (workspace_id) REFERENCES workspace(id)
